test-component/requirements.txt
pika
paho-mqtt
#matplotlib
influxdb